Refactor BlockingKernelClient/AsyncKernelClient
