+++ setup.py
@@ -74,6 +74,7 @@ setup_args = dict(
         'pyzmq>=13',
         'python-dateutil>=2.1',
         'tornado>=4.1',
+        'nest-asyncio>=1.5',
     ],
     python_requires  = '>=3.5',
     extras_require   = {
